Move out DividedConstructor visitor
